+++ src/main/java/me/lkp111138/dealbot/commands/HelpCommand.java
@@ -11,7 +11,7 @@ import me.lkp111138.dealbot.translation.Translation;
 public class HelpCommand implements Command {
     @Override
     public void respond(TelegramBot bot, Message msg, String[] args) {
-        int sender = msg.from().id();
+        long sender = msg.from().id();
         if (msg.chat().type() == Chat.Type.Private) {
             SendMessage req = new SendMessage(sender, Translation.get(DealBot.lang(sender)).HELP()).parseMode(ParseMode.HTML);
             bot.execute(req);
