+++ src/main/java/me/lkp111138/dealbot/commands/HowToCommand.java
@@ -11,7 +11,7 @@ import me.lkp111138.dealbot.translation.Translation;
 public class HowToCommand implements Command {
     @Override
     public void respond(TelegramBot bot, Message msg, String[] args) {
-        int sender = msg.from().id();
+        long sender = msg.from().id();
         if (msg.chat().type() == Chat.Type.Private) {
             SendMessage req = new SendMessage(sender, Translation.get(DealBot.lang(sender)).HOW_TO_PLAY()).parseMode(ParseMode.HTML);
             bot.execute(req);
