+++ src/main/java/me/lkp111138/dealbot/commands/JoinCommand.java
@@ -21,12 +21,12 @@ public class JoinCommand implements Command {
         try (Connection conn = Main.getConnection()) {
             // create user profile if there wasn't one
             PreparedStatement stmt = conn.prepareStatement("INSERT IGNORE INTO tg_users (tgid, username) VALUES (?, ?)");
-            stmt.setInt(1, msg.from().id());
+            stmt.setLong(1, msg.from().id());
             stmt.setString(2, msg.from().username());
             stmt.execute();
             stmt.close();
             stmt = conn.prepareStatement("DELETE FROM next_game WHERE tgid=? AND gid=?");
-            stmt.setInt(1, msg.from().id());
+            stmt.setLong(1, msg.from().id());
             stmt.setLong(2, msg.chat().id());
             stmt.execute();
             stmt.close();
